update to `1.0.1` added : - `langful.replace` function - tests for `langful.replace` function changes : - change the `langful.merge` function - change the `default.LANG`

+++ src/langful/loader.py
@@ -44,6 +44,7 @@ class loader :
         with open( file , "rb" ) as fp : return self.suffixes[ suffix ].load( fp.read() )
 
     def save( self , file : str , data : dict[ str , typing.Any ] , suffix : str | None = None ) -> None :
+        if os.path.exists( file ) and not os.path.isfile( file ) : raise IsADirectoryError( "the path is not a file" )
         if suffix is None : suffix = os.path.splitext( file )[ -1 ]
         if suffix not in self : raise KeyError( f"no parser can load file with '{ suffix }' suffix" )
         with open( file , "wb" ) as fp : fp.write( self.suffixes[ suffix ].save( data ) )
